+++ HISTORY.md
@@ -18,6 +18,7 @@
 ====================
   * Added method KuzFile.ok()
   * Now checking if ok() before passing code / kuz to page
+  * Added KuzFile.getCachedItems()
 
 
 0.2.6 / 27-09-2020